made hw-auth

+++ src/controllers/notesController.js
@@ -6,7 +6,9 @@ export const getAllNotes = async (req, res) => {
   const skip = (page - 1) * perPage;
   const limit = perPage;
 
-  const notesQuery = Note.find();
+  const notesQuery = Note.find({
+    userId: req.user._id,
+  });
   if (search) {
     notesQuery.where({
       $or: [
@@ -38,32 +40,32 @@ const [totalNotes, notes] = await Promise.all([
 
 export const getNoteById = async (req, res) => {
   const { noteId } = req.params;
-  const note = await Note.findById(noteId);
+  const note = await Note.findOne({ _id: noteId, userId: req.user._id });
   if (!note) {
-    throw createHttpError(404, `Note with ID: ${noteId} not found`);
+    throw createHttpError(404, `Note not found`);
   }
   res.status(200).json(note);
 };
 
 export const createNote = async(req, res) => {
- const newNote = await Note.create(req.body);
+ const newNote = await Note.create({...req.body, userId: req.user._id});
   res.status(201).json(newNote);
 };
 
 export const deleteNote = async (req, res) => {
   const { noteId } = req.params;
-  const deletedNote = await Note.findOneAndDelete({ _id: noteId });
+  const deletedNote = await Note.findOneAndDelete({ _id: noteId, userId: req.user._id });
   if (!deletedNote) {
-    throw createHttpError(404, `Note with ID: ${noteId} not found`);
+    throw createHttpError(404, `Note not found`);
   }
   res.status(200).json(deletedNote);
 };
 
 export const updateNote = async (req, res) => {
   const { noteId } = req.params;
-  const updatedNote = await Note.findOneAndUpdate({ _id: noteId }, req.body, { returnDocument: 'after' });
+  const updatedNote = await Note.findOneAndUpdate({ _id: noteId, userId: req.user._id }, req.body, { returnDocument: 'after' });
   if (!updatedNote) {
-    throw createHttpError(404, `Note with ID: ${noteId} not found`);
+    throw createHttpError(404, `Note not found`);
   }
   res.status(200).json(updatedNote);
 };
